Update OBDZero repository to version 4.15
Version 4.15 targets Android 15 however it ops out of edge to edge on on Android 15. This means that it will not work with Android 16 where edge to edge displays are mandatory. OBDZero supplies information on range, speed and battery condition for use while driving the iMiev/CZero/iOn series of cars. To avoid distracting the driver, the display is as simple as possible. Edge to Edge defeats this purpose.